feat: Abstract date formatting functions into the date utilities file
