+++ lib/screens/events_browser/events_screen.dart
@@ -32,6 +32,7 @@ import 'package:bluecherry_client/screens/events_browser/filter.dart';
 import 'package:bluecherry_client/screens/events_browser/sidebar.dart';
 import 'package:bluecherry_client/screens/players/event_player_desktop.dart';
 import 'package:bluecherry_client/utils/constants.dart';
+import 'package:bluecherry_client/utils/date.dart';
 import 'package:bluecherry_client/utils/extensions.dart';
 import 'package:bluecherry_client/utils/methods.dart';
 import 'package:bluecherry_client/widgets/desktop_buttons.dart';
